Movie Rental API/requirements.txt
fastapi==0.95.2
uvicorn==0.23.1
asyncpg==0.29.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
passlib==1.7.4
jose==3.3.0
